requirements.txt
numpy
Pillow
torch
keras
tensorflow
matplotlib
Juypyter